backend/requirements.txt
fastapi
uvicorn[standard]
requests
pyyaml # 用于Clash的YAML处理
aiohttp # 异步HTTP请求
httpx # 异步HTTP客户端，requests的现代替代品